requirements.txt
psycopg2
python-dotenv==0.15.0
pandas
